add params_good option for SaveAll, more boa cleanup and move config to separate go file
