+++ examples/boa/configs/2csperdrive.toml
@@ -1,0 +1,5 @@
+[Env]
+  CSPerDrive = 2
+  PctCortexStEpc = 20
+
+
